refactor: extract shared handleTabListKeyDown into uiUtils
